fix(macos,widgets,settings): mac toolbar+cards, Today widget bg, extract Other settings page
- MacClassTableView: replace Label with explicit HStack(Image, Text)+fixedSize
  so the toolbar add-course glyph and title stay on the same baseline and
  AppKit can't squeeze them into icon-only when the semester picker widens.
- TodayWidget: paint palette.background as an explicit ZStack bottom layer
  in addition to containerBackground so the homescreen surface has a real
  color even if WidgetKit skips the container background on this family.
- MacWidgetCards: today's row gets the iPhone-style course-color treatment
  (live row solid, upcoming rows tinted strip + 10% fill) and the next
  class label is now a colored chip so the Mac cards read as colorful too.
- SettingsView: pull invert-slider through view-source-code into a new
  OtherSettingsView sub-page reached from a NavigationLink under the
  existing Other settings header; drop the now-unused state, sheets, alert
  and reassignAllCourseColors helper.
