+++ swift/TigerDuck/Features/Settings/SettingsView.swift
@@ -9,9 +9,6 @@ struct SettingsView: View {
     @State private var notifyFreeLunch = true
     @State private var notifyClubs = false
     @State private var showingTabEditor = false
-    @State private var showLicense = false
-    @State private var showPrivacyPolicy = false
-    @State private var showFeedback = false
     @State private var showLibraryLogin = false
     @State private var libIsLoggingIn = false
     @State private var libLoginError: String?
@@ -22,13 +19,9 @@ struct SettingsView: View {
     @State private var hapticEngine: CHHapticEngine?
     @State private var hapticPlayer: CHHapticPatternPlayer?
     @State private var notificationsAuthorized: Bool = true
-    @State private var showReassignColorsConfirm = false
     @State private var showOfficialWebsite = false
     @Environment(\.scenePhase) private var scenePhase
 
-    private static let feedbackURL = AppURLs.issues
-    private static let privacyURL = AppURLs.privacyPolicy
-    private static let licenseURL = AppURLs.license
     private static let deleteAccountURL = AppURLs.deleteAccount
     private static let websiteURL = AppURLs.website
 
@@ -124,57 +117,13 @@ struct SettingsView: View {
             }
 
             // MARK: - Other settings
-            // One header'd subsection with the Library toggle, then a string of
-            // header-less Sections so each row (or grouping) gets its own card
-            // — mirrors the Android `OtherSettingsScreen` layout where every
-            // ContentCard is visually separated. iOS Form can't nest Sections,
-            // so the visual grouping is achieved with multiple top-level
-            // Sections sharing the same conceptual "Other settings" header.
+            // Library toggle keeps its position at the top of the
+            // "Other settings" group; the rest of the miscellany now lives
+            // behind a NavigationLink to `OtherSettingsView`.
             Section(String(localized: "settings_section_other_settings")) {
                 Toggle(String(localized: "settings_library_related_features"), isOn: libraryToggleBinding)
-            }
-            Section {
-                Toggle(String(localized: "settings_invert_slider_direction"), isOn: $appState.invertSliderDirection)
-            }
-            Section {
-                Button {
-                    showReassignColorsConfirm = true
-                } label: {
-                    Text(String(localized: "settings_reset_course_colors"))
-                        .foregroundStyle(.primary)
-                }
-            }
-            Section {
-                Button {
-                    if appState.browserPreference == .inApp {
-                        showFeedback = true
-                    } else {
-                        UIApplication.shared.open(Self.feedbackURL)
-                    }
-                } label: {
-                    Text(String(localized: "settings_feedback_bug_report"))
-                        .foregroundStyle(.primary)
-                }
-                Button {
-                    if appState.browserPreference == .inApp {
-                        showPrivacyPolicy = true
-                    } else {
-                        UIApplication.shared.open(Self.privacyURL)
-                    }
-                } label: {
-                    Text(String(localized: "settings_privacy_policy"))
-                        .foregroundStyle(.primary)
-                }
-                Button(String(localized: "settings_open_source_licenses")) {
-                    if appState.browserPreference == .inApp {
-                        showLicense = true
-                    } else {
-                        UIApplication.shared.open(Self.licenseURL)
-                    }
-                }
-                .foregroundStyle(.primary)
-                NavigationLink(String(localized: "settings_view_source_code")) {
-                    SourceCodePickerView()
+                NavigationLink(String(localized: "settings_section_other_settings")) {
+                    OtherSettingsView()
                 }
             }
 
@@ -269,32 +218,9 @@ struct SettingsView: View {
         .sheet(isPresented: $showingTabEditor) {
             TabEditorView()
         }
-        .sheet(isPresented: $showFeedback) {
-            InAppBrowserView(url: Self.feedbackURL)
-                .ignoresSafeArea()
-        }
-        .sheet(isPresented: $showPrivacyPolicy) {
-            InAppBrowserView(url: Self.privacyURL)
-                .ignoresSafeArea()
-        }
-        .sheet(isPresented: $showLicense) {
-            InAppBrowserView(url: Self.licenseURL)
-                .ignoresSafeArea()
-        }
         .sheet(isPresented: $showOfficialWebsite) {
             InAppBrowserView(url: Self.websiteURL)
                 .ignoresSafeArea()
-        }
-        .alert(
-            String(localized: "settings_reset_course_colors_confirm_title"),
-            isPresented: $showReassignColorsConfirm
-        ) {
-            Button(String(localized: "action_confirm"), role: .destructive) {
-                reassignAllCourseColors()
-            }
-            Button(String(localized: "action_cancel"), role: .cancel) {}
-        } message: {
-            Text(String(localized: "settings_reset_course_colors_confirm_message"))
         }
         .sheet(isPresented: $showLibraryLogin) {
             LoginSheet(
@@ -370,17 +296,6 @@ struct SettingsView: View {
             libraryWarningTask?.cancel()
             libraryWarningTask = nil
         }
-    }
-
-    /// Rebuild every course's color assignment from scratch using the
-    /// unique-color algorithm, then broadcast so Home, Class Table, widgets,
-    /// and the Live Activity all pick up the new palette. Reads the
-    /// authoritative roster from the canonical course provider so user-added
-    /// courses participate too.
-    private func reassignAllCourseColors() {
-        let courseNos = CanonicalCourseProvider().currentCourses().map(\.courseNo)
-        TigerDuckTheme.reassignAll(courseNos: courseNos)
-        NotificationCenter.default.post(name: AppConstants.dataDidUpdate, object: nil)
     }
 
     private var libraryToggleBinding: Binding<Bool> {
